requirements.txt
groq>=0.11.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
duckduckgo-search>=6.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
rich>=13.7.0
